Move `FeedImageDataLoaderPresentationAdapter` to its own file
